% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_kpi.R
\name{compute_kpis}
\alias{compute_kpis}
\title{Compute all KPIs}
\usage{
compute_kpis(data)
}
\arguments{
\item{data}{A data frame containing the necessary columns.}
}
\value{
A list of all computed KPIs.
}
\description{
Compute all KPIs
}
