% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter.R
\name{.filter}
\alias{.filter}
\title{Filter a dataset}
\usage{
.filter(data, arglist)
}
\arguments{
\item{data}{the data frame}

\item{arglist}{a list of filter specifications}
}
\description{
A function to filter a data frame based on specified list.
The list should contain elements with the following structure:
- col: the name of the column to filter on
- fun: the filtering function to use ("between" or "in")
- min, max: the range for "between" filters
- val: the values for "in" filters

Modified from Teofil Nakov's blog:
https://discindo.org/posts/2025-01-31-filter-snippet/
}
\examples{
d <- iris
l <- list(
  list(col = "Sepal.Length", fun = "between", min = 4, max = 5),
  list(col = "Species", fun = "in", val = c("setosa", "versicolor"))
)
.filter(d, l)
}
