% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample dataset
A sample dataset containing various measurements.}
\format{
A data frame with 3000 rows and 6 columns:
\describe{
  \item{id}{Unique identifier for each observation}
  \item{site}{Location where the measurement was taken}
  \item{date}{Date of the measurement}
  \item{type}{Type of measurement}
  \item{value}{Value of the measurement}
  \item{carbon.emission.in.kgco2e}{Carbon emission in kgCO2e}
}
}
\usage{
sample_data
}
\description{
Sample dataset
A sample dataset containing various measurements.
}
\keyword{datasets}
