% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_validate.R
\name{validate_dataset}
\alias{validate_dataset}
\title{validate_dataset}
\usage{
validate_dataset(input_data = NULL)
}
\arguments{
\item{input_data}{A data.frame to validate.}
}
\value{
Either TRUE or an error message.
}
\description{
A function to validate the dataset.
}
\examples{
validate_dataset(sample_data)
}
